Add Gas Benchmarking Tests for WebAuthn Signer
